+++ src/api.jl
@@ -170,6 +170,15 @@ function get_oauth_client(ctx::Context, id::AbstractString; kw...)
     return _get(ctx, joinpath(PATH_OAUTH_CLIENTS, id); kw...).client
 end
 
+# Returns the user with the given email.
+function find_user(ctx::Context, email::AbstractString; kw...)
+    rsp = list_users(ctx)
+    for item in rsp
+        item["email"] == email && return item
+    end
+    return nothing
+end
+
 function get_user(ctx::Context, userid::AbstractString; kw...)
     return _get(ctx, joinpath(PATH_USERS, userid); kw...).user
 end