+++ CHANGELOG.md
@@ -1,2 +1,7 @@
 # Changelog
 
+## main
+
+* add find_user to api.jl
+* rename examples/get-userid.jl to examples/find-user.jl
+